requirements.txt
numpy
pyaudio
rpi_ws281x
paho-mqtt
